StateDraft to interface. #942
